댓글 CommentContent 컴포넌트로 따로 뺌
