+++ interfaces/board.interface.ts
@@ -1,24 +1,24 @@
-import { User } from "./user.interface";
+import { User } from './user.interface';
 
 export interface Post {
-    _id: string;
-    author: User;
-    contents: string;
-    images: string[];
-    comments: Comment[];
-    createdAt: Date;
-    updatedAt: Date;
-    likes: User[];
-    category: string;
-  }
+  _id: string;
+  author: User;
+  contents: string;
+  images: string[];
+  comments: Comment[];
+  createdAt: Date;
+  updatedAt: Date;
+  likes: User[];
+  category: string;
+}
 
 export interface Comment {
-    _id: string;
-    user_id: User;
-    createdAt: Date;
-    content: string;
-    likes: User[];
-    is_deleted: boolean;
+  _id: string;
+  author: User;
+  createdAt: Date;
+  content: string;
+  likes: User[];
+  is_deleted: boolean;
 }
 
 export interface boardContents {
